requirements.txt
asgiref==3.6.0
Django==4.2
psycopg2-binary==2.9.6
sqlparse==0.4.4
django-crispy-forms==1.14.0
crispy-bootstrap5==0.6
django-widget-tweaks
python-docx==0.8.11
selenium~=4.9.0